backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
pyjwt
python-jose
email-validator
google-auth
google-auth-oauthlib
requests
pydantic-settings
python-dotenv
